Remove stateful logic from the Creator utilities
